refactor: use appState file to keep all the recoil and hooks functions
