--- src/hooks/useMerkleDistributor.ts
@@ -1,6 +1,0 @@
-import { useRecoilValue } from 'recoil';
-
-import { rMerkleDistributor } from 'recoilState/merkleDistributor';
-
-export const useMerkleDistributor = (epochId: number) =>
-  useRecoilValue(rMerkleDistributor(epochId));
