--- src/hooks/useReceivedGiftsByUser.ts
@@ -1,6 +1,0 @@
-import { useRecoilValue } from 'recoil';
-
-import { rReceivedGiftsByUser } from 'recoilState/receivedGiftsByUser';
-
-export const useReceivedGiftsByUser = (epochId: number) =>
-  useRecoilValue(rReceivedGiftsByUser(epochId));
