+++ src/recoilState/appState.ts
@@ -1,4 +1,8 @@
 // at 5k elements for filter-map-slice itiriri is more performant
+import {
+  MerkleDistributorInfo,
+  parseBalanceMap,
+} from 'helpers/merkle-distributor/parse-balance-map';
 import iti from 'itiriri';
 import moment from 'moment';
 import {
@@ -18,6 +22,7 @@ import {
   createExtendedEpoch,
 } from 'utils/modelExtenders';
 import storage from 'utils/storage';
+import { assertDef } from 'utils/tools';
 
 import { rMyAddress } from './walletState';
 
@@ -500,3 +505,56 @@ export const rUserGifts = selectorFamily<
 });
 export const useUserGifts = (userId: number) =>
   useRecoilValue(rUserGifts(userId));
+
+export type IReceivedGiftsByUser = { [account: string]: number | string };
+
+export const rReceivedGiftsByUser = selectorFamily<
+  IReceivedGiftsByUser,
+  number
+>({
+  key: 'rReceivedGiftsByUser',
+  get: (epochId: number) => ({ get }: IRecoilGetParams) => {
+    const usersMap = get(rUsersMap);
+    const gifts = get(rGifts).filter((g) => g.epoch_id === epochId);
+    const giftsByUser = iti(gifts).toGroups(
+      (g) => g.recipient_id,
+      (g) => g.tokens
+    );
+    const totalReceivedByUserMap = iti(giftsByUser.keys()).toMap(
+      (userId) =>
+        assertDef(
+          usersMap.get(userId)?.address,
+          'Missing user while calculating receivedGiftsByUser'
+        ),
+      (userId) => iti(assertDef(giftsByUser.get(userId))).sum() ?? 0
+    );
+
+    const totalReceivedByUser: IReceivedGiftsByUser = {};
+    totalReceivedByUserMap.forEach((val, key) => {
+      if (val > 0) {
+        totalReceivedByUser[key] = val;
+      }
+    });
+
+    return totalReceivedByUser;
+  },
+});
+
+export const useReceivedGiftsByUser = (epochId: number) =>
+  useRecoilValue(rReceivedGiftsByUser(epochId));
+
+export const rEpochMerkleRoot = selectorFamily<
+  MerkleDistributorInfo | undefined,
+  number
+>({
+  key: 'rEpochMerkleRoot',
+  get: (epochId: number) => ({ get }: IRecoilGetParams) => {
+    const receivedGiftsByUser = get(rReceivedGiftsByUser(epochId));
+    if (Object.keys(receivedGiftsByUser).length !== 0) {
+      return parseBalanceMap(receivedGiftsByUser);
+    }
+  },
+});
+
+export const useEpochMerkleRoot = (epochId: number) =>
+  useRecoilValue(rEpochMerkleRoot(epochId));
